Migrated ?ranks and ?list commands, improved exception-handling, added a global channel check, created base for leaderboards command class
