+++ Common.py
@@ -1,0 +1,20 @@
+"""
+Contains common code
+"""
+
+# list of channels in which Sneasel can trigger commands
+channel_list = ["leaderboards", "sneasel_commands", "support"]
+
+# list of all leaderboards
+leaderboard_list = ['jogger', 'pikachu', 'battlegirl', 'pokedex', 'collector', 'scientist', 'breeder', 'backpacker',
+                    'fisherman', 'youngster', 'berrymaster', 'gymleader', 'champion', 'battlelegend', 'ranger', 'unown',
+                    'gentleman', 'pilot', 'totalxp', 'goldgyms', 'idol', 'greatleague', 'ultraleague', 'masterleague',
+                    'acetrainer', 'cameraman', 'hero', 'purifier']
+
+# list of channels that correspond to leaderboards
+leaderboard_channels = []
+
+# Developers: McMomo,
+developers = [169688623699066880]
+
+current_leaderboard = ""
