--- Includes.py
@@ -1,15 +1,0 @@
-"""
-Contains common code
-"""
-
-# list of channels in which Sneasel can trigger commands
-channelList = ["leaderboards, sneasel_commands, support"]
-
-# list of all leaderboards
-leaderboard_list = ["jogger", "pikachu", "battlegirl", "pokedex", "collector", "scientist", "breeder", "backpacker", "fisherman",
-                        "youngster", "berrymaster", "gymleader", "champion", "battlelegend", "ranger", "unown", "gentleman",
-                        "pilot", "totalxp", "goldgyms", "idol", "greatleague", "ultraleague", "masterleague", "acetrainer",
-                        "cameraman", "hero", "purifier"]
-
-# Developers: McMomo
-developers = [169688623699066880]
